Preparing for controls update.
I'm re-branding this as "Beam Plays a Keyboard". I want to avoid code churn
and limit forks. If you want to configure a game. You shouldn't have to touch
code.

We've swapped back to defaulting to robotjs because it works for literally
every other game other than the sega collection. To not churn the code too
much the keyboard handler is now configureable in the config file.

I intend to write config files for many popular games and provide instructions
on how to set one up.