Added basic blocking system.
Most consoles and games have a key reset combo that involves start and select
as these buttons are usually separate in a game we can lock them out. This
process isn't configureable right now but can be made so.

Depending on the mode set in the upcoming "mode" feature we can also block out
left and right from being used etc. This will reduce the ingame character
wobbling back and forth.