Update config.
I have 5 config files with different controls. To fix my password being
everywhere. I've now allowed for an auth.json file that allows you to
keep it in one place, As a bonus it lets me upload my keyConfigs for games
as I make them. Old config files are backwards compatible.

We also now listen to the threshold as defined in the config.

Blocks can be specified in the config as "blocks : { "start":"select"} etc.