Validate beam controls, throw errors correctly, reduce complexity.
If an error occured we were picking it up in our sigint/exit handler and
clearing keys. this called the widgets which clears the console. This no
longer happens so that when an error occurs we log it to the console correctly
as an error.

Most process.exit calls have been replaced with errors, this will allow
for upcoming modularization of this project.

I ran some static code analysis on this project and it pointed out some areas
of repetition. This project was a hacky quick attempt at a beam interactive game.
but now its almost required by the beam community. Its therefore gotta stop
being hacky and start being slick and smooth.